% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsne.R
\name{tsne}
\alias{tsne}
\title{t-Distributed Stochastic Neighbor Embedding}
\usage{
tsne(data, no_dims=2, perplexity=30, theta=0.5, eta=200, exageration=12,
            iterations=1000, verbose=FALSE)
}
\arguments{
\item{data}{A matrix}

\item{no_dims}{Output dimensionality}

\item{perplexity}{Perplexity parameter}

\item{theta}{Speed vs accuracy trade-off}

\item{eta}{Learning rate}

\item{exageration}{exageration parameter}

\item{iterations}{Number of iterations}

\item{verbose}{Whether progress updates should be printed}
}
\value{
A matrix containing the new representations for the data
}
\description{
Wrapper for a CUDA implementation of Barnes-Hut t-Distributed Stochastic Neighbor Embedding (t-SNE).
t-SNE is a technique for dimensionality reduction that is particularly well suited for the
visualization of high-dimensional datasets.
}
\references{
Maaten, L. Van Der, 2014. Accelerating t-SNE using Tree-Based Algorithms.
            Journal of Machine Learning Research, 15, p.3221-3245.

            van der Maaten, L.J.P. & Hinton, G.E., 2008. Visualizing
            High-Dimensional Data Using t-SNE. Journal of Machine Learning
            Research, 9, pp.2579-2605.
}
